Move DisposableToolTip to EditorFeatures.Wpf
